refactor(frontend): split vibe/ytmusic/youtube/tidal/listen-groups api mixins
Complete the mechanical domain split: move the remaining vibe, YouTube Music,
YouTube, Tidal, and listen-group domain methods into typed mixin modules.
api.ts is now a thin facade (imports, shared response types, mixin
composition, singleton) reduced from 3741 to 387 lines across 23 cohesive
domain modules plus core. Pure structural move.

Refactor slice P34 (api-client-typing-split).
